Some syntax fixes

+++ src/com/bs/model/Department.java
@@ -1,16 +1,16 @@
 package com.bs.model;
 
 public class Department {
-	Integer id;
+	long id;
 	String department_name;
 
-	public Department(Integer id, String department_name) {
+	public Department(long id, String department_name) {
 
 		this.id = id;
 		this.department_name = department_name;
 	}
 
-	public Integer getId() {
+	public long getId() {
 		return id;
 	}
